issue #34: LRP toolbox now runs under Pythoin 3. Use requirements_py36_conda.txt ot env_lrp_py36.yaml to recreate the lrp_py36 environments

+++ Software/lrp_toolbox-1.2.0/python/modules/maxpool.py
@@ -61,8 +61,8 @@ class MaxPool(Module):
         hstride, wstride = self.stride
 
         #assume the given pooling and stride parameters are carefully chosen.
-        Hout = (H - hpool) / hstride + 1
-        Wout = (W - wpool) / wstride + 1
+        Hout = int((H - hpool) / hstride + 1)
+        Wout = int((W - wpool) / wstride + 1)
 
         #initialize pooled output
         self.Y = np.zeros((N,Hout,Wout,D))
@@ -102,14 +102,14 @@ class MaxPool(Module):
         hstride, wstride = self.stride
 
         #assume the given pooling and stride parameters are carefully chosen.
-        Hout = (H - hpool) / hstride + 1
-        Wout = (W - wpool) / wstride + 1
+        Hout = int((H - hpool) / hstride + 1)
+        Wout = int((W - wpool) / wstride + 1)
 
         #distribute the gradient towards the max activation(s)
         #the max activation value is already known via self.Y
         DX = np.zeros_like(self.X,dtype=np.float)
-        for i in range(Hout):
-            for j in range(Wout):
+        for i in range(int(Hout)):
+            for j in range(int(Wout)):
                 DX[:,i*hstride:i*hstride+hpool , j*wstride:j*wstride+wpool,:] += DY[:,i:i+1,j:j+1,:] * (self.Y[:,i:i+1,j:j+1,:] == self.X[:, i*hstride:i*hstride+hpool , j*wstride:j*wstride+wpool , : ])
         return DX
 
@@ -166,13 +166,13 @@ class MaxPool(Module):
         hstride, wstride = self.stride
 
         #assume the given pooling and stride parameters are carefully chosen.
-        Hout = (H - hpool) / hstride + 1
-        Wout = (W - wpool) / wstride + 1
+        Hout = int((H - hpool) / hstride + 1)
+        Wout = int((W - wpool) / wstride + 1)
 
         Rx = np.zeros_like(self.X,dtype=np.float)
 
-        for i in range(Hout):
-            for j in range(Wout):
+        for i in range(int(Hout)):
+            for j in range(int(Wout)):
                 Z = self.Y[:,i:i+1,j:j+1,:] == self.X[:, i*hstride:i*hstride+hpool , j*wstride:j*wstride+wpool , : ]
                 Zs = Z.sum(axis=(1,2),keepdims=True)
                 Rx[:,i*hstride:i*hstride+hpool , j*wstride:j*wstride+wpool,:] += (Z / Zs) * R[:,i:i+1,j:j+1,:]
@@ -189,8 +189,8 @@ class MaxPool(Module):
         hstride, wstride = self.stride
 
         #assume the given pooling and stride parameters are carefully chosen.
-        Hout = (H - hpool) / hstride + 1
-        Wout = (W - wpool) / wstride + 1
+        Hout = int((H - hpool) / hstride + 1)
+        Wout = int((W - wpool) / wstride + 1)
 
         Rx = np.zeros_like(self.X,dtype=np.float)
